Split editor into multiple component files
