+++ app/components/ProseEditor/BlockStyleControls.js
@@ -1,0 +1,41 @@
+import React from 'react';
+import StyleButton from './StyleButton';
+
+import styles from './ProseEditor.css';
+
+const BLOCK_TYPES = [
+  { label: 'H1', style: 'header-one' },
+  { label: 'H2', style: 'header-two' },
+  { label: 'H3', style: 'header-three' },
+  { label: 'H4', style: 'header-four' },
+  { label: 'H5', style: 'header-five' },
+  { label: 'H6', style: 'header-six' },
+  { label: 'Blockquote', style: 'blockquote' },
+  { label: 'UL', style: 'unordered-list-item' },
+  { label: 'OL', style: 'ordered-list-item' },
+  { label: 'Code Block', style: 'code-block' }
+];
+
+const BlockStyleControls = props => {
+  const { editorState, onToggle } = props;
+  const selection = editorState.getSelection();
+  const blockType = editorState
+    .getCurrentContent()
+    .getBlockForKey(selection.getStartKey())
+    .getType();
+  return (
+    <div className={styles.controls}>
+      {BLOCK_TYPES.map(type => (
+        <StyleButton
+          key={type.label}
+          active={type.style === blockType}
+          label={type.label}
+          onToggle={onToggle}
+          style={type.style}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default BlockStyleControls;
