+++ app/components/ProseEditor/InlineStyleControls.js
@@ -1,0 +1,32 @@
+import React from 'react';
+import StyleButton from './StyleButton';
+
+import styles from './ProseEditor.css';
+
+const INLINE_STYLES = [
+  { label: 'Bold', style: 'BOLD' },
+  { label: 'Italic', style: 'ITALIC' },
+  { label: 'Underline', style: 'UNDERLINE' },
+  { label: 'Monospace', style: 'CODE' }
+];
+
+const InlineStyleControls = props => {
+  const { editorState, onToggle } = props;
+  const currentStyle = editorState.getCurrentInlineStyle();
+
+  return (
+    <div className={styles.controls}>
+      {INLINE_STYLES.map(type => (
+        <StyleButton
+          key={type.label}
+          active={currentStyle.has(type.style)}
+          label={type.label}
+          onToggle={onToggle}
+          style={type.style}
+        />
+      ))}
+    </div>
+  );
+};
+
+export default InlineStyleControls;
