+++ app/components/ProseEditor/ProseEditor.css
@@ -1,0 +1,65 @@
+/**
+ * Adapted from draft-js rich editor example
+ * https://github.com/facebook/draft-js/blob/master/examples/draft-0-10-0/rich/
+ */
+
+.root {
+  background: #fff;
+  font-size: 14px;
+  padding: 15px;
+}
+
+.editor {
+  cursor: text;
+  font-size: 16px;
+  margin-top: 10px;
+}
+
+.editor :global(.public-DraftEditorPlaceholder-root),
+.editor :global(.public-DraftEditor-contents) {
+  margin: 0 -15px -15px;
+  padding: 15px;
+}
+
+.editor :global(.public-DraftEditor-content) {
+  min-height: 100px;
+}
+
+.hidePlaceholder :global(.public-DraftEditorPlaceholder-root) {
+  display: none;
+}
+
+.editor .blockquote {
+  border-left: 5px solid #eee;
+  color: #666;
+  font-family: "Hoefler Text", "Georgia", serif;
+  font-style: italic;
+  margin: 16px 0;
+  padding: 10px 20px;
+}
+
+.editor :global(.public-DraftStyleDefault-pre) {
+  background-color: rgba(0, 0, 0, 0.05);
+  font-family: "Inconsolata", "Menlo", "Consolas", monospace;
+  font-size: 16px;
+  padding: 20px;
+}
+
+.controls {
+  /* font-family: "Helvetica", sans-serif; */
+  font-size: 14px;
+  margin-bottom: 5px;
+  user-select: none;
+}
+
+.styleButton {
+  color: #999;
+  cursor: pointer;
+  margin-right: 16px;
+  padding: 2px 0;
+  display: inline-block;
+}
+
+.activeButton {
+  color: #5890ff;
+}
