+++ app/components/ProseEditor/index.js
@@ -1,0 +1,1 @@
+export { default as ProseEditor } from './Editor';
